+++ src/cli.py
@@ -1,7 +1,8 @@
 import argparse
 import random
+from typing import Union
 
-from .core import DEFAULT_THREADS, SCHEDULER_FORK_SCALE, SCHEDULER_INITIAL_CAPACITY
+from .core import COPIES_AUTO, DEFAULT_THREADS, SCHEDULER_FORK_SCALE, SCHEDULER_INITIAL_CAPACITY
 from .i18n import LANGUAGES
 from .mhddos import Methods
 
@@ -26,9 +27,9 @@ def init_argparse() -> argparse.ArgumentParser:
     )
     parser.add_argument(
         '--copies',
-        type=int,
+        type=lambda val: val if val == COPIES_AUTO else int(val),
         default=1,
-        help='Number of copies (default is 1)',
+        help='Number of copies (default is 1). Use "auto" to set the value automatically',
     )
     parser.add_argument(
         '--debug',