+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+- **08.06.2020**
+  - Added `--copies auto` option to set the value automatically based on the resources available
+
+- **25.05.2022**
+  - Improved default output - the `--debug` option is no longer required 
+
 - **21.05.2022**
   - Added english localization - use flag `--lang EN` (more languages may be added later)